requirements.txt
easyocr
Flask
numpy
opencv-python
Pillow
ultralytics